Add the "-time <n>" flag to limit the first client wait duration
If tty-share is executed on-demand to serve a console to a
remote user, we want to limit the duration the tty server will
wait for the client to connect.

If the requester doesn't use the tty he requested, leaving the
server running is both a security liability (brute force url
forging attack) and a undue resource usage (port, memory, cpu).